refactor: rename UserProfile to ProfileMenu and update related imports
